Improved ClassAvailability view and refactored all classes packages
