+++ app/src/main/AndroidManifest.xml
@@ -12,7 +12,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".MainActivity"
+            android:name=".Activities.MainActivity"
             android:screenOrientation="portrait"
             android:theme="@style/generalNoTitle">
             <intent-filter>
@@ -22,7 +22,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".LoginActivity"
+            android:name=".Activities.LoginActivity"
             android:label="@string/app_name"
             android:screenOrientation="portrait">
             <intent-filter>
@@ -37,16 +37,16 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".MainMenuActivity"
+            android:name=".Activities.MainMenuActivity"
             android:label=""
             android:screenOrientation="portrait"
             android:theme="@style/MainMenuTheme" />
         <activity
-            android:name=".AboutActivity"
+            android:name=".Activities.AboutActivity"
             android:theme="@style/generalNoTitle" />
-        <activity android:name=".NotificationDetailsActivity" />
+        <activity android:name=".Activities.NotificationDetailsActivity" />
         <activity
-            android:name=".SettingsActivity"
+            android:name=".Activities.SettingsActivity"
             android:screenOrientation="portrait"/>
     </application>
