+++ app/src/main/res/layout/activity_about.xml
@@ -9,7 +9,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.upc.fib.racopocket.AboutActivity">
+    tools:context=".Activities.AboutActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
