+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/raco_blue"
-    tools:context="com.upc.fib.racopocket.MainActivity">
+    tools:context=".Activities.MainActivity">
 
     <ImageView
         android:layout_width="wrap_content"
