+++ app/src/main/res/layout/activity_notification_details.xml
@@ -7,7 +7,7 @@
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
-    tools:context=".NotificationDetailsActivity">
+    tools:context=".Activities.NotificationDetailsActivity">
 
     <TextView
         android:id="@+id/notification_title"
