+++ app/src/main/res/layout/app_bar_main_menu.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="com.upc.fib.racopocket.MainMenuActivity">
+    tools:context=".Activities.MainMenuActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
