+++ app/src/main/res/layout/class_availability_item_list.xml
@@ -12,7 +12,6 @@
     <TextView android:id="@+id/availabilityClassAvailability"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
-        android:layout_weight=".4"
-        android:gravity="center"/>
+        android:layout_weight=".4"/>
 
 </LinearLayout>
