+++ app/src/main/res/values-ca/strings.xml
@@ -91,8 +91,8 @@
     <string name="update_data_title">Descarregant les dades necessàries</string>
     <string name="update_data_description">Actualitza totes les dades de l\'aplicació (Recomanat cada cop que comença un curs nou).</string>
     <string name="update_data_text">Totes les dades s\'actualitzaran, i això pot consumir uns quants MB d\'internet. Et recomanem activar la connexió Wi-fi.</string>
-    <string name="update_data_successful">Totes les dades s'han actualitzat correctament</string>
-    <string name="update_data_unsuccessful">Algo ha fallat, comprova la teva connexió i torna a intentar-ho</string>
+    <string name="update_data_successful">Totes les dades s\'han actualitzat correctament</string>
+    <string name="update_data_unsuccessful">Alguna cosa ha fallat, comprova la teva connexió i torna a intentar-ho</string>
     <string name="continue_dialog">Continuar</string>
     <string name="cancel_dialog">Cancel·lar</string>
